feat(backend): removed canvas-maker service + small refactoring
