--- scripts/helmcharts/openreplay/charts/canvas-maker/templates/service.yaml
@@ -1,18 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ include "canvas-maker.fullname" . }}
-  namespace: {{ .Release.Namespace }}
-  labels:
-    {{- include "canvas-maker.labels" . | nindent 4 }}
-spec:
-  type: {{ .Values.service.type }}
-  ports:
-    {{- range $key, $val := .Values.service.ports }}
-    - port: {{ $val }}
-      targetPort: {{ $key }}
-      protocol: TCP
-      name: {{ $key }}
-    {{- end}}
-  selector:
-    {{- include "canvas-maker.selectorLabels" . | nindent 4 }}
