--- scripts/helmcharts/openreplay/charts/canvas-maker/templates/serviceMonitor.yaml
@@ -1,18 +1,0 @@
-{{- if and ( .Capabilities.APIVersions.Has "monitoring.coreos.com/v1" ) ( .Values.serviceMonitor.enabled ) }}
-apiVersion: monitoring.coreos.com/v1
-kind: ServiceMonitor
-metadata:
-  name: {{ include "canvas-maker.fullname" . }}
-  namespace: {{ .Release.Namespace }}
-  labels:
-    {{- include "canvas-maker.labels" . | nindent 4 }}
-  {{- if .Values.serviceMonitor.additionalLabels }}
-    {{- toYaml .Values.serviceMonitor.additionalLabels | nindent 4 }}
-  {{- end }}
-spec:
-  endpoints:
-    {{- .Values.serviceMonitor.scrapeConfigs | toYaml | nindent 4 }}
-  selector:
-    matchLabels:
-      {{- include "canvas-maker.selectorLabels" . | nindent 6 }}
-{{- end }}
