--- scripts/helmcharts/openreplay/charts/canvas-maker/templates/tests/test-connection.yaml
@@ -1,15 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: "{{ include "canvas-maker.fullname" . }}-test-connection"
-  labels:
-    {{- include "canvas-maker.labels" . | nindent 4 }}
-  annotations:
-    "helm.sh/hook": test
-spec:
-  containers:
-    - name: wget
-      image: busybox
-      command: ['wget']
-      args: ['{{ include "canvas-maker.fullname" . }}:{{ .Values.service.port }}']
-  restartPolicy: Never
