--- scripts/helmcharts/openreplay/charts/canvas-maker/values.yaml
@@ -1,124 +1,0 @@
-# Default values for openreplay.
-# This is a YAML-formatted file.
-# Declare variables to be passed into your templates.
-
-replicaCount: 1
-
-image:
-  repository: "{{ .Values.global.openReplayContainerRegistry }}/canvas-maker"
-  pullPolicy: IfNotPresent
-  # Overrides the image tag whose default is the chart appVersion.
-  tag: ""
-
-imagePullSecrets: []
-nameOverride: "canvas-maker"
-fullnameOverride: "canvas-maker-openreplay"
-
-serviceAccount:
-  # Specifies whether a service account should be created
-  create: true
-  # Annotations to add to the service account
-  annotations: {}
-  # The name of the service account to use.
-  # If not set and create is true, a name is generated using the fullname template
-  name: ""
-
-podAnnotations: {}
-
-securityContext:
-  runAsUser: 1001
-  runAsGroup: 1001
-podSecurityContext:
-  runAsUser: 1001
-  runAsGroup: 1001
-  fsGroup: 1001
-  fsGroupChangePolicy: "OnRootMismatch"
-
-#securityContext: {}
-  # capabilities:
-  #   drop:
-  #   - ALL
-  # readOnlyRootFilesystem: true
-  # runAsNonRoot: true
-  # runAsUser: 1000
-
-service:
-  type: ClusterIP
-  ports:
-    http: 9000 
-    metrics: 8888
-
-serviceMonitor:
-  enabled: true
-  additionalLabels:
-    release: observability
-  scrapeConfigs:
-    - port: metrics
-      honorLabels: true
-      interval: 15s
-      path: /metrics
-      scheme: http
-      scrapeTimeout: 10s
-
-ingress:
-  enabled: false
-  className: ""
-  annotations: {}
-    # kubernetes.io/ingress.class: nginx
-    # kubernetes.io/tls-acme: "true"
-  hosts:
-    - host: chart-example.local
-      paths:
-        - path: /
-          pathType: ImplementationSpecific
-  tls: []
-  #  - secretName: chart-example-tls
-  #    hosts:
-  #      - chart-example.local
-
-resources: {}
-  # We usually recommend not to specify default resources and to leave this as a conscious
-  # choice for the user. This also increases chances charts run on environments with little
-  # resources, such as Minikube. If you do want to specify resources, uncomment the following
-  # lines, adjust them as necessary, and remove the curly braces after 'resources:'.
-  # limits:
-  #   cpu: 100m
-  #   memory: 128Mi
-  # requests:
-  #   cpu: 100m
-  #   memory: 128Mi
-
-autoscaling:
-  enabled: false
-  minReplicas: 1
-  maxReplicas: 5
-  targetCPUUtilizationPercentage: 80
-  # targetMemoryUtilizationPercentage: 80
-
-env:
-  FS_CLEAN_HRS: 24
-
-pvc:
-  # This can be either persistentVolumeClaim or hostPath.
-  # In case of pvc, you'll have to provide the pvc name.
-  # For example
-  # name: openreplay-efs
-  name: "{{ .Values.global.pvcRWXName }}"
-  hostMountPath: /openreplay/storage/nfs
-
-persistence: {}
-  # # Spec of spec.template.spec.containers[*].volumeMounts
-  # mounts:
-  # - name: kafka-ssl
-  #   mountPath: /opt/kafka/ssl
-  # # Spec of spec.template.spec.volumes
-  # volumes:
-  # - name: kafka-ssl
-  #   secret:
-  #     secretName: kafka-ssl
-
-nodeSelector: {}
-
-tolerations: []
-
-affinity: {}
